// src/types/transactions.ts
import type { Category, CategorySummary } from "./category";

export enum TransactionType {
    INCOME = 'income',
    EXPENSE = 'expense',
}

export interface Transaction {
    id: string;
    userId: string;
    description: string;
    amount: number;
    date: string | Date;
    type: TransactionType;
    categoryId: string;
    category: Category;
    updatedAt: string | Date;
    createdAt: string | Date;
}


export interface TransactionFilter {
    month: number;
    year: number;
    categoryId: string;
    type?: TransactionType;
}

export interface TransactionSummary {
  totalExpenses: number;        
  totalIncomes: number;         
  balance: number;              
  previousBalance: number;      
  monthResult: number;         
  expensesByCategory: CategorySummary[];
}

export interface MonthlyItem {
    name: string;
    exoenses: number;
    incomes: number;

} 

export interface CreateTransactionDTO {
  description: string;
  amount: number;
  date: Date | string;
  categoryId: string;
  type: TransactionType;
}